move `RollingBallShaker` and `RollingBallStopper` to `GameObjects`
